fix: Sections 일급클래스 기반으로 메서드 정의 -> SectionController와 SectionService 삭제
